pyproject.toml
[project]
name = "senacfinal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-admin-interface>=0.30.0",
    "django-allauth>=65.7.0",
    "django-cors-headers>=4.7.0",
    "django-role-permissions>=3.2.0",
    "pillow>=11.2.1",
]

[dependency-groups]
dev = [
    "python-decouple>=3.8",
]

[tool.taskipy.tasks]
server = "uv run manage.py runserver"
make = "uv run manage.py makemigrations"
migrate = "uv run manage.py migrate"
create-admin = "uv run manage.py createsuperuser"
